renamed selection to drag. Created new file called selectionmodel. Added a new class GhostSelectionModel with a modified select method for ghost columns.
