+++ package.json
@@ -57,7 +57,7 @@
     "@lumino/widgets": "^1.13.3",
     "@types/codemirror": "0.0.76",
     "tde-csvviewer": "^1.1.3",
-    "tde-datagrid": "^1.0.7",
+    "tde-datagrid": "^1.0.8",
     "tde-datastore": "^0.1.1",
     "tde-dragdrop": "^0.0.1"
   },
